♻️(frontend) move default view into demo
Instead of having the default views in
the package and like the demo will be public.
It is better to put the default views in the demo.
