+++ src/frontend/demo/jest-setup.ts
@@ -1,1 +1,10 @@
-import "@testing-library/jest-dom";
+import '@testing-library/jest-dom';
+import { server } from './src/mocks/server';
+
+// Establish API mocking before all tests.
+beforeAll(() => server.listen());
+// Reset any request handlers that we may add during the tests,
+// so they don't affect other tests.
+afterEach(() => server.resetHandlers());
+// Clean up after the tests are finished.
+afterAll(() => server.close());
