+++ src/frontend/demo/package.json
@@ -43,6 +43,7 @@
     "jest": "28.1.1",
     "jest-css-modules": "2.1.0",
     "jest-environment-jsdom": "28.1.1",
+    "msw": "0.47.4",
     "prettier": "2.7.0",
     "ts-jest": "28.0.4",
     "typescript": "4.7.3"
@@ -52,7 +53,7 @@
     "extract-translations": "formatjs extract './**/*.ts*' --ignore ./node_modules --ignore './**/*.d.ts' --ignore './**/*.stories.ts*' --out-file './i18n/frontend.json' --id-interpolation-pattern '[sha512:contenthash:base64:6]' --format crowdin",
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "jest",
+    "test": "TZ=UTC jest",
     "eject": "react-scripts eject",
     "lint": "eslint 'src/**/*.{ts,tsx}'",
     "format:check": "prettier --check src",
