+++ src/frontend/demo/src/mocks/handlers/index.ts
@@ -1,0 +1,4 @@
+import { loginHandlers } from './login/loginHandlers';
+import { roomsHandlers } from './rooms/roomsHandlers';
+
+export const handlers = [...loginHandlers, ...roomsHandlers];
