+++ src/frontend/demo/src/utils/routes/rooms/index.tsx
@@ -1,6 +1,7 @@
-import { RoomSettingsView, RoomsListView } from '@jitsi-magnify/core';
 import { defineMessages, IntlShape } from 'react-intl';
 import { Link, RouteObject } from 'react-router-dom';
+import { RoomsListView } from '../../../views/rooms/list';
+import { RoomSettingsView } from '../../../views/rooms/settings';
 
 export enum RoomsPath {
   ROOMS = '/app/rooms',
