+++ src/frontend/demo/src/views/auth/register/index.tsx
@@ -1,0 +1,33 @@
+import { AuthForms, SignupForm } from '@jitsi-magnify/core';
+import * as React from 'react';
+import { defineMessages, useIntl } from 'react-intl';
+import { useNavigate } from 'react-router-dom';
+import { AuthPath } from '../../../utils/routes/auth';
+
+const authRegisterViewMessages = defineMessages({
+  youAlreadyHaveAnAccount: {
+    defaultMessage: 'You already have an account?',
+    description: 'The explanation for the link to the login form',
+    id: 'views.auth.register.youAlreadyHaveAnAccount',
+  },
+  loginInstead: {
+    defaultMessage: 'Login instead',
+    description: 'The text for the link to the login form',
+    id: 'views.auth.register.loginInstead',
+  },
+});
+
+export function AuthRegisterView() {
+  const intl = useIntl();
+  const navigate = useNavigate();
+  return (
+    <AuthForms
+      footerLabel={intl.formatMessage(authRegisterViewMessages.youAlreadyHaveAnAccount)}
+      footerRedirect={() => navigate(AuthPath.LOGIN)}
+      footerRouteLabel={intl.formatMessage(authRegisterViewMessages.loginInstead)}
+      isLogin={true}
+    >
+      <SignupForm />
+    </AuthForms>
+  );
+}
