+++ src/frontend/magnify/src/index.ts
@@ -3,5 +3,6 @@ export * from './i18n';
 export * from './themes';
 export * from './middleware';
 export * from './context';
-export * from './utils/routes/api';
-export * from './views';
+export * from './utils';
+export * from './services';
+export * from './types';
