+++ src/frontend/magnify/src/mocks/allHandlers/rooms/roomsHandlers.ts
