+++ src/frontend/magnify/src/services/http/index.ts
@@ -1,0 +1,1 @@
+export * from './http.service';
