+++ src/frontend/magnify/src/services/index.ts
@@ -1,2 +1,4 @@
 export * from './rooms';
 export * from './users';
+export * from './routes';
+export * from './http';
