+++ src/frontend/magnify/src/services/rooms/index.ts
@@ -1,0 +1,1 @@
+export * from './rooms.repository';
