+++ src/frontend/magnify/src/services/routes/index.ts
@@ -1,0 +1,1 @@
+export * from './RoutesBuilder.service';
