+++ src/frontend/magnify/src/types/api/index.ts
@@ -1,0 +1,3 @@
+export * from './auth';
+export { LoginResponse } from './auth';
+export * from './room';
