+++ src/frontend/magnify/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './entities';
+export * from './api';
