--- src/frontend/magnify/src/utils/routes/api/users.routes.ts
@@ -1,9 +1,0 @@
-export enum UsersRoutes {
-  REFRESH_TOKEN = '/accounts/token-refresh/',
-  ME = '/users/me/',
-  LOGIN = '/users/login/',
-  CHANGE_PASSWORD = '/users/change-password/',
-  UPDATE = '/users/:id/',
-  DELETE = '/users/:id/',
-  CREATE = '/users/',
-}
