+++ src/frontend/magnify/src/utils/routes/api/users/index.ts
@@ -1,0 +1,1 @@
+export * from './usersApiRoutes';
