+++ src/frontend/magnify/src/utils/routes/api/users/usersApiRoutes.ts
@@ -1,0 +1,9 @@
+export enum UsersApiRoutes {
+  REFRESH_TOKEN = '/accounts/token-refresh/',
+  ME = '/users/me/',
+  LOGIN = '/users/login/',
+  CHANGE_PASSWORD = '/users/change-password/',
+  UPDATE = '/users/:id/',
+  DELETE = '/users/:id/',
+  CREATE = '/users/',
+}
