--- src/frontend/magnify/src/views/auth/index.ts
@@ -1,2 +1,0 @@
-export * from './login';
-export * from './register';
