--- src/frontend/magnify/src/views/index.ts
@@ -1,3 +1,0 @@
-export * from './rooms';
-export * from './auth';
-export * from './profile';
