--- src/frontend/magnify/src/views/profile/index.ts
@@ -1,1 +1,0 @@
-export * from './account';
