--- src/frontend/magnify/src/views/rooms/index.ts
@@ -1,2 +1,0 @@
-export * from './list';
-export * from './settings';
